% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_scrape.r
\name{enhanced_PBP}
\alias{enhanced_PBP}
\title{Enhance NHL RTSS Data}
\usage{
enhanced_PBP(rawdata)
}
\arguments{
\item{rawdata}{NHL RTSS data}
}
\description{
Create additional values for each line of NHL RTSS Data
}
\details{
Included Items \cr
\itemize{
  \item is_home = 1 if the home team was the primary focus of the event
  \item game_mins = Converts game seconds to minutes
  \item side_coordsx and side_coordsy = Moves all event coordinates to one side on the ice based on is_home
  \item is_corsi = i if the event is a corsi event (Shot, Goal, Miss, or Block)
  \item event_angle = the angle (in degrees) of a given event to the goal
  \item event_distance = the distance (in feet) of a given event to the goal
}
}
\author{
Eric Fastner (eric.fastner@gmail.com)
}
