% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EH_scrape.r
\name{expand_pbp}
\alias{expand_pbp}
\title{Add additional data to RTSS data}
\usage{
expand_pbp(data)
}
\arguments{
\item{data}{pbp_base data frame created by scrape_pbp()}
}
\description{
Add additional data to RTSS data
}
\author{
Evolving Hockey (https://evolving-hockey.com/)
}
