% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_summary.r
\name{get_game_results}
\alias{get_game_results}
\alias{get_team_results}
\title{Summarize Game Results}
\usage{
get_game_results(raw_data, full_season = FALSE)

get_team_results(raw_data, full_season = FALSE)
}
\arguments{
\item{raw_data}{A data frame containing NHL RTSS data}

\item{full_season}{if TRUE will add columns to summarize total number of games played as well as total points earned}
}
\description{
Utilize NHL RTSS game data to summarize game results. \cr
\cr
Use get_game_results() to summarize by game ID \cr
Use get_team_results() to summarize by team
}
\author{
Eric Fastner (eric.fastner@gmail.com)
}
