% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalie_summary.r
\name{get_goalie_stats}
\alias{get_goalie_stats}
\title{Summarize goalie stats by game ID}
\usage{
get_goalie_stats(dataset)
}
\description{
Create a data frame of all goalies contained in an NHL RTSS data frame
}
\details{
\itemize{
   \item TOI = Time on Ice
   \item SA = Shots Against
   \item GA = Goals Against
   \item SV = Saves
   \item SV_Perc = Save Percentage
   \item GS = \href{https://hockey-graphs.com/2016/07/13/measuring-single-game-productivity-an-introduction-to-game-score/}{Gamescore}
}
}
\author{
Eric Fastner (eric.fastner@gmail.com)
}
