% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icescrapr.r
\docType{package}
\name{icescrapr}
\alias{icescrapr}
\alias{icescrapr-package}
\title{icescrapR: A package for helping to scrape, summarize, and visualize the NHL RTSS (play-by-play) data}
\description{
A collection of functions that will help the user srape
    play-by-play data from NHL.com as well as a collection of functions
    for summarizing, analyzing, and visualizing the results
}
\section{Scrape Functions}{

enhanced_PBP \cr
}

\section{Summarize Functions}{

get_game_results \cr
get_team_results \cr
}

\section{Stats Functions}{

get_skater_stats \cr
}

\section{Visualization Functions}{

viz_draw_rink \cr
viz_corsi_graph \cr
viz_corsi_positions \cr
viz_team_summary \cr
}

\section{Data Sets}{

alignment_20002011 \cr
alignment_20112013 \cr
alignment_present \cr
corsi_events \cr
gs_weights_skaters \cr
gs_weights_goalies \cr
rink \cr
team_colors \cr
}

