% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EH_scrape.r
\name{scrape_pbp}
\alias{scrape_pbp}
\title{Scrape NHL RTSS (play-by-play) data from nhl.com}
\usage{
scrape_pbp(games, scrape_type = "full", live_scrape = FALSE,
  verbose = TRUE, sleep = 0)
}
\arguments{
\item{games}{a vector containing a list of game_ids to scrape}

\item{scrape_type}{default is "full", determines what data to return. See details for more info}

\item{live_scrape}{default is FALSE, adjusts incorrect player and goalie shifts if TRUE (used for scraping games in progress)}

\item{verbose}{default is TRUE, prints system time of each scrape}

\item{sleep}{default is 0, time (in seconds) to wait between scraping games}
}
\description{
Scrape NHL RTSS (play-by-play) data from nhl.com
}
\details{
\strong{\emph{event_summary values}}
\itemize{
   \item \strong{full} - returns all data
   \item \strong{event_summary} - returns only event summary, rosters, and scratches
   \item \strong{rosters} - returns only rosters and scratches
}

\strong{\emph{resulting data sets returned}}
\itemize{
   \item \strong{game_info_dir} - game information data
   \item \strong{pbp_base} - main play-by-play data
   \item \strong{pbp_extras} - extra play-by-play data
   \item \strong{player_shifts} - full player shifts data
   \item \strong{player_periods} - player TOI sums per period (from the shifts source)
   \item \strong{roster_df} - roster data
   \item \strong{scratches_df} - scratches data
   \item \strong{events_summary_df} - event summary data (box score stats, etc.)
   \item \strong{report} - report showing number of rows and time to scrape game
}
}
\author{
Evolving Hockey (https://evolving-hockey.com/)
}
