% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EH_scrape.r
\name{scrape_player_info}
\alias{scrape_player_info}
\title{Scrape all player info for a given season}
\usage{
scrape_player_info(season_id_fun)
}
\arguments{
\item{season_id_fun}{an 8 digit value created by combining the start year and end year of the season (ie '20172018')}
}
\description{
Scrape all player info for a given season
}
\author{
(Evolving Hockey https://evolving-hockey.com/)
}
