% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EH_scrape.r
\name{scrape_schedule}
\alias{scrape_schedule}
\title{Scrape a list of games played during a given date interval}
\usage{
scrape_schedule(start_date, end_date, print_sched = TRUE)
}
\arguments{
\item{start_date}{the desired starting date}

\item{end_date}{the desired end date}

\item{print_sched}{default is TRUE, prints resulting dataframe to console}
}
\description{
Scrape a list of games played during a given date interval
}
\author{
Evolving Hockey (https://evolving-hockey.com/)
}
