% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_viz.r
\name{viz_corsi_graph}
\alias{viz_corsi_graph}
\title{Chart Corsi Events for Each Team}
\usage{
viz_corsi_graph(rawdata, color_list = team_colors)
}
\arguments{
\item{rawdata}{NHL RTSS data for a single game}

\item{color_list}{utilizes the hex colors stored in team_colors$primary unless otherwise specified}
}
\description{
Create a chart of all corsi events for each team plotted by minute of
the game using NHL RTSS data. Only uses one game at a time.
}
\author{
Eric Fastner (eric.fastner@gmail.com)
}
