% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_viz.r
\name{viz_corsi_positions}
\alias{viz_corsi_positions}
\title{Chart Corsi Events and Their Positions}
\usage{
viz_corsi_positions(rawdata, color_list = team_colors)
}
\arguments{
\item{rawdata}{NHL RTSS data for a single game}

\item{color_list}{utilizes the hex colors stored in team_colors$primary unless otherwise specified}
}
\description{
Create a visual representation of all corsi events for a given game and their position on the ice
}
\author{
Eric Fastner (eric.fastner@gmail.com)
}
