% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_viz.r
\name{viz_draw_rink}
\alias{viz_draw_rink}
\title{Visual Representation of Hockey Rink}
\usage{
viz_draw_rink()
}
\description{
Create a visual representation of a hockey rink for use with ggplot2
}
\author{
Prashanth Iyer (Twitter: \href{https://twitter.com/iyer_prashanth}{@iyer_prashanth})
}
