% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_viz.r
\name{viz_team_summary}
\alias{viz_team_summary}
\title{Graphically Summarize Team Stats}
\usage{
viz_team_summary(rawdata, color_list = team_colors)
}
\description{
Create a graphical representation of specific team stats for a given game
}
\details{
\itemize{
  \item Blocks = Total blocks by each team
  \item Hits = Total hits by each team
  \item Faceoffs = Total faceoffs won by each team
  \item Corsi = Total corsi events by each team
  \item Shots = Total shots by each team
  \item Goals = Total goals by each team
}
}
\author{
Eric Fastner (eric.fastner@gmail.com)
}
